added togglinng in blog card (#3370)
