+++ src/components/About/Blog/index.js
@@ -1,137 +1,16 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import htmlToText from 'html-to-text';
 import { connect } from 'react-redux';
-import _Card from '@material-ui/core/Card';
-import _CardMedia from '@material-ui/core/CardMedia';
-import dateFormat from 'dateformat';
 import Fab from '@material-ui/core/Fab';
-import {
-  FacebookShareButton,
-  TwitterShareButton,
-  TelegramShareButton,
-  WhatsappShareButton,
-  FacebookIcon,
-  TwitterIcon,
-  TelegramIcon,
-  WhatsappIcon,
-  LinkedinShareButton,
-  LinkedinIcon,
-} from 'react-share';
-import styled, { css } from 'styled-components';
-import renderHTML from 'react-render-html';
+import styled from 'styled-components';
 import { getBlogReponse } from '../../../apis';
 import './Blog.css';
 import 'font-awesome/css/font-awesome.min.css';
 import Next from '@material-ui/icons/KeyboardArrowRight';
 import Previous from '@material-ui/icons/KeyboardArrowLeft';
-import susi from '../../../images/susi-logo.svg';
 import { Header } from '../../shared/About';
-
 import BlogLoader from './BlogLoader';
-const allCategories = [
-  'FOSSASIA',
-  'GSoC',
-  'SUSI.AI',
-  'Tutorial',
-  'Android',
-  'API',
-  'App generator',
-  'CodeHeat',
-  'Community',
-  'Event',
-  'Event Management',
-  'loklak',
-  'Meilix',
-  'Open Event',
-  'Phimpme',
-  'Pocket Science Lab',
-  'yaydoc',
-];
-
-const arrDiff = (a1, a2) => {
-  let a = [],
-    diff = [];
-  for (let f = 0; f < a1.length; f++) {
-    a[a1[f]] = true;
-  }
-  for (let z = 0; z < a2.length; z++) {
-    if (a[a2[z]]) {
-      delete a[a2[z]];
-    } else {
-      a[a2[z]] = true;
-    }
-  }
-  for (let k in a) {
-    diff.push(k);
-  }
-  return diff;
-};
-
-const LinkStyle = css`
-  font-size: 0.875rem;
-  text-decoration: none;
-  color: rgba(255, 255, 255, 0.54);
-`;
-
-const FlexBox = styled.div`
-  display: flex;
-  align-items: baseline;
-  color: rgba(51, 51, 51, 0.7);
-  line-height: 1.563rem;
-  margin-right: 2rem;
-`;
-
-const BlogPostContainer = styled.div`
-  padding: 1rem;
-`;
-
-const BlogFooter = styled.div`
-  padding: 3rem;
-  background: #f7f7f7;
-  display: flex;
-  flex-wrap: wrap;
-`;
-
-const CardMedia = styled(_CardMedia)`
-  height: 0;
-  padding-top: 56.25%;
-  height: 31.25rem;
-  object-fit: contain;
-  vertical-align: middle;
-
-  @media (max-width: 1000px) {
-    height: 18.75rem;
-  }
-`;
-
-const Card = styled(_Card)`
-  position: relative;
-`;
-
-const Overlay = styled.div`
-  position: relative;
-  left: 0rem;
-  background: rgba(0, 0, 0, 0.54);
-  width: 100%;
-  padding: 1rem;
-  margin-top: -3.5rem;
-`;
-
-const Heading = styled.p`
-  font-size: 2.125rem;
-  font-family: 'Roboto', 'Helvetica', 'Arial', sans-serif;
-  font-weight: 400;
-  line-height: 1.1;
-  margin: 0;
-  letter-spacing: 0.00735em;
-`;
-
-const CustomTypography = styled.div`
-  margin-bottom: 2rem;
-  color: rgba(0, 0, 0, 0.54);
-  font-size: 0.875rem;
-`;
+import BlogItem from './BlogItem';
 
 const BlogNavigation = styled.div`
   justify-content: center;
@@ -148,43 +27,8 @@ const BlogNavigation = styled.div`
   }
 `;
 
-const SocialButtons = styled.div`
-  width: 100%;
-  display: flex;
-  padding: 0.625rem 0 1.25rem 0.625rem;
-  align-items: center;
-  justify-content: flex-start;
-  cursor: pointer;
-  * {
-    margin-right: 5px;
-  }
-`;
-
-const Icon = styled.i`
-  padding-right: 0.625rem;
-`;
-
 const Container = styled.div`
   width: 100%;
-`;
-
-const BlogFooterLink = styled.a`
-  white-space: nowrap;
-  cursor: pointer;
-  text-decoration: none;
-  color: rgba(0, 0, 0, 0.87);
-  height: 1.875rem;
-
-  &:hover {
-    text-decoration: underline;
-    color: rgba(51, 51, 51, 0.7);
-  }
-`;
-
-const OverlayLink = styled.a`
-  &&& {
-    ${LinkStyle};
-  }
 `;
 
 const BottomPost = styled.div`
@@ -316,139 +160,9 @@ class Blog extends Component {
               {this.state.posts &&
                 Array.isArray(this.state.posts) &&
                 this.state.posts.length > 0 &&
-                this.state.posts.map((posts, i) => {
-                  const description = htmlToText
-                    .fromString(posts.description)
-                    .split('…');
-                  const content = posts.content;
-                  let category = [];
-                  posts.categories.forEach(cat => {
-                    let k = 0;
-                    for (k = 0; k < allCategories.length; k++) {
-                      if (cat === allCategories[k]) {
-                        category.push(cat);
-                      }
-                    }
-                  });
-
-                  const tags = arrDiff(category, posts.categories);
-                  const fCategory = category
-                    .map((cat, index) => (
-                      <BlogFooterLink
-                        key={index}
-                        href={
-                          'http://blog.fossasia.org/category/' +
-                          cat.replace(/\s+/g, '-').toLowerCase()
-                        }
-                        rel="noopener noreferrer"
-                      >
-                        {cat}
-                      </BlogFooterLink>
-                    ))
-                    .reduce((prev, curr) => [prev, ', ', curr]);
-                  const ftags = tags
-                    .map((tag, index) => (
-                      <BlogFooterLink
-                        key={index}
-                        href={
-                          'http://blog.fossasia.org/tag/' +
-                          tag.replace(/\s+/g, '-').toLowerCase()
-                        }
-                        rel="noopener noreferrer"
-                      >
-                        {tag}
-                      </BlogFooterLink>
-                    ))
-                    .reduce((prev, curr) => [prev, ', ', curr]);
-                  let htmlContent = content.replace(/<img.*?>/, '');
-                  htmlContent = renderHTML(htmlContent);
-                  let image = susi;
-                  const regExp = /\[(.*?)\]/;
-                  const imageUrl = regExp.exec(description[0]);
-                  if (imageUrl) {
-                    image = imageUrl[1];
-                  }
-                  const date = posts.pubDate.split(' ');
-                  const d = new Date(date[0]);
-                  return (
-                    <div key={i} className="section_blog">
-                      <Card>
-                        <CardMedia image={image} />
-                        <Overlay>
-                          <OverlayLink href={posts.link}>
-                            {`Published on ${dateFormat(
-                              d,
-                              'dddd, mmmm dS, yyyy',
-                            )}`}
-                          </OverlayLink>
-                        </Overlay>
-                        <BlogPostContainer>
-                          <Heading>{posts.title}</Heading>
-                          <CustomTypography>
-                            by
-                            <a
-                              style={{ paddingLeft: '0.3rem' }}
-                              href={`http://blog.fossasia.org/author/${posts.author}`}
-                            >
-                              {posts.author}
-                            </a>
-                          </CustomTypography>
-                          <div>{htmlContent}</div>
-                        </BlogPostContainer>
-                        <SocialButtons>
-                          <TwitterShareButton
-                            url={posts.guid}
-                            title={posts.title}
-                            via="asksusi"
-                            hashtags={posts.categories.slice(0, 4)}
-                          >
-                            <TwitterIcon size={32} round={true} />
-                          </TwitterShareButton>
-                          <FacebookShareButton url={posts.link}>
-                            <FacebookIcon size={32} round={true} />
-                          </FacebookShareButton>
-                          <LinkedinShareButton url={posts.link}>
-                            <LinkedinIcon size={32} round={true} />
-                          </LinkedinShareButton>{' '}
-                          <WhatsappShareButton url={posts.link}>
-                            <WhatsappIcon size={32} round={true} />
-                          </WhatsappShareButton>
-                          <TelegramShareButton url={posts.link}>
-                            <TelegramIcon size={32} round={true} />
-                          </TelegramShareButton>
-                        </SocialButtons>
-                        <BlogFooter>
-                          <FlexBox>
-                            <Icon className="fa fa-calendar" />
-                            <BlogFooterLink href={posts.link}>
-                              {dateFormat(d, 'mmmm dd, yyyy')}
-                            </BlogFooterLink>
-                          </FlexBox>
-                          <FlexBox>
-                            <Icon className="fa fa-user" />
-                            <BlogFooterLink
-                              rel="noopener noreferrer"
-                              href={
-                                'http://blog.fossasia.org/author/' +
-                                posts.author
-                              }
-                            >
-                              {posts.author}
-                            </BlogFooterLink>
-                          </FlexBox>
-                          <FlexBox>
-                            <Icon className="fa fa-folder-open-o" />
-                            {fCategory}
-                          </FlexBox>
-                          <FlexBox>
-                            <Icon className="fa fa-tags" />
-                            <div>{ftags}</div>
-                          </FlexBox>
-                        </BlogFooter>
-                      </Card>
-                    </div>
-                  );
-                })}
+                this.state.posts.map((posts, i) => (
+                  <BlogItem posts={posts} key={i} />
+                ))}
             </Container>
             <BlogNavigation>
               <Fab
